use coil for image loading

+++ app/src/main/java/com/example/theguide/presentation/dashboard/DashboardScreen.kt
@@ -16,14 +16,18 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import androidx.lifecycle.viewmodel.compose.viewModel
 import com.example.theguide.R
 import com.example.theguide.domain.model.Place
 import com.example.theguide.presentation.navigation.Route
+import com.example.theguide.presentation.welcome.WelcomeVM
 import com.example.theguide.ui.component.PrimaryTopAppBar
 import com.example.theguide.ui.component.RecommendationCard
 import com.example.theguide.ui.theme.TheGuideTheme
@@ -56,7 +60,7 @@ fun DashboardScreen(
                 ) {
                     items(state.places) { place ->
                         RecommendationCard(
-                            image = place.imageUrl.toInt(),
+                            image = place.imageUrl,
                             name = place.name,
                             rating = place.rating.toString(),
                             onClick = {  }
